+++ src/JsonExporterTest.php
@@ -61,7 +61,7 @@ final class JsonExporterTest extends TestCase
         $handle = fopen($this->tmpFile, 'wb');
         $exporter = new JsonExporter($handle, 0);
         $data = Statistics::fromOne(Unit::Nanoseconds, 3);
-        $exporter->exportStatistics($data);
+        $exporter->exportStatistics($data, 'test');
         fclose($handle);
 
         /** @var string $json */
@@ -69,7 +69,7 @@ final class JsonExporterTest extends TestCase
         self::assertJson($json);
 
         /** @var non-empty-string $jsonData */
-        $jsonData = json_encode($data);
+        $jsonData = json_encode([...['label' => 'test'], ...$data->toHuman()]);
         self::assertSame($jsonData, $json);
     }